// src/rnode-protos.d.ts
export interface IPar {

    /** Par sends */
    sends?: (ISend[]|null);

    /** Par receives */
    receives?: (IReceive[]|null);

    /** Par news */
    news?: (INew[]|null);

    /** Par exprs */
    exprs?: (IExpr[]|null);

    /** Par matches */
    matches?: (IMatch[]|null);

    /** Par unforgeables */
    unforgeables?: (IGUnforgeable[]|null);

    /** Par bundles */
    bundles?: (IBundle[]|null);

    /** Par connectives */
    connectives?: (IConnective[]|null);

    /** Par locallyFree */
    locallyFree?: (Uint8Array|null);

    /** Par connectiveUsed */
    connectiveUsed?: (boolean|null);
}
// ...
export interface IBundle {

    /** Bundle body */
    body?: (IPar|null);

    /** Bundle writeFlag */
    writeFlag?: (boolean|null);

    /** Bundle readFlag */
    readFlag?: (boolean|null);
}
// ...
export interface ISend {

    /** Send chan */
    chan?: (IPar|null);

    /** Send data */
    data?: (IPar[]|null);

    /** Send persistent */
    persistent?: (boolean|null);

    /** Send locallyFree */
    locallyFree?: (Uint8Array|null);

    /** Send connectiveUsed */
    connectiveUsed?: (boolean|null);
}
// ...
export interface IReceiveBind {

    /** ReceiveBind patterns */
    patterns?: (IPar[]|null);

    /** ReceiveBind source */
    source?: (IPar|null);

    /** ReceiveBind remainder */
    remainder?: (IVar|null);

    /** ReceiveBind freeCount */
    freeCount?: (number|null);
}
// ...
export interface IReceive {

    /** Receive binds */
    binds?: (IReceiveBind[]|null);

    /** Receive body */
    body?: (IPar|null);

    /** Receive persistent */
    persistent?: (boolean|null);

    /** Receive peek */
    peek?: (boolean|null);

    /** Receive bindCount */
    bindCount?: (number|null);

    /** Receive locallyFree */
    locallyFree?: (Uint8Array|null);

    /** Receive connectiveUsed */
    connectiveUsed?: (boolean|null);
}
// ...
export interface INew {

    /** New bindCount */
    bindCount?: (number|null);

    /** New p */
    p?: (IPar|null);

    /** New uri */
    uri?: (string[]|null);

    /** New deployId */
    deployId?: (IDeployId|null);

    /** New deployerId */
    deployerId?: (IDeployerId|null);

    /** New locallyFree */
    locallyFree?: (Uint8Array|null);
}
// ...
export interface IMatch {

    /** Match target */
    target?: (IPar|null);

    /** Match cases */
    cases?: (IMatchCase[]|null);

    /** Match locallyFree */
    locallyFree?: (Uint8Array|null);

    /** Match connectiveUsed */
    connectiveUsed?: (boolean|null);
}
// ...
export interface IExpr {

    /** Expr gBool */
    gBool?: (boolean|null);

    /** Expr gInt */
    gInt?: (number|Long|null);

    /** Expr gString */
    gString?: (string|null);

    /** Expr gUri */
    gUri?: (string|null);

    /** Expr gByteArray */
    gByteArray?: (Uint8Array|null);

    /** Expr eNotBody */
    eNotBody?: (IENot|null);

    /** Expr eNegBody */
    eNegBody?: (IENeg|null);

    /** Expr eMultBody */
    eMultBody?: (IEMult|null);

    /** Expr eDivBody */
    eDivBody?: (IEDiv|null);

    /** Expr ePlusBody */
    ePlusBody?: (IEPlus|null);

    /** Expr eMinusBody */
    eMinusBody?: (IEMinus|null);

    /** Expr eLtBody */
    eLtBody?: (IELt|null);

    /** Expr eLteBody */
    eLteBody?: (IELte|null);

    /** Expr eGtBody */
    eGtBody?: (IEGt|null);

    /** Expr eGteBody */
    eGteBody?: (IEGte|null);

    /** Expr eEqBody */
    eEqBody?: (IEEq|null);

    /** Expr eNeqBody */
    eNeqBody?: (IENeq|null);

    /** Expr eAndBody */
    eAndBody?: (IEAnd|null);

    /** Expr eOrBody */
    eOrBody?: (IEOr|null);

    /** Expr eVarBody */
    eVarBody?: (IEVar|null);

    /** Expr eListBody */
    eListBody?: (IEList|null);

    /** Expr eTupleBody */
    eTupleBody?: (IETuple|null);

    /** Expr eSetBody */
    eSetBody?: (IESet|null);

    /** Expr eMapBody */
    eMapBody?: (IEMap|null);

    /** Expr eMethodBody */
    eMethodBody?: (IEMethod|null);

    /** Expr eMatchesBody */
    eMatchesBody?: (IEMatches|null);

    /** Expr ePercentPercentBody */
    ePercentPercentBody?: (IEPercentPercent|null);

    /** Expr ePlusPlusBody */
    ePlusPlusBody?: (IEPlusPlus|null);

    /** Expr eMinusMinusBody */
    eMinusMinusBody?: (IEMinusMinus|null);

    /** Expr eModBody */
    eModBody?: (IEMod|null);
}
// ...
export interface IEList {

    /** EList ps */
    ps?: (IPar[]|null);

    /** EList locallyFree */
    locallyFree?: (Uint8Array|null);

    /** EList connectiveUsed */
    connectiveUsed?: (boolean|null);

    /** EList remainder */
    remainder?: (IVar|null);
}
// ...
export interface IETuple {

    /** ETuple ps */
    ps?: (IPar[]|null);

    /** ETuple locallyFree */
    locallyFree?: (Uint8Array|null);

    /** ETuple connectiveUsed */
    connectiveUsed?: (boolean|null);
}
// ...
export interface IESet {

    /** ESet ps */
    ps?: (IPar[]|null);

    /** ESet locallyFree */
    locallyFree?: (Uint8Array|null);

    /** ESet connectiveUsed */
    connectiveUsed?: (boolean|null);

    /** ESet remainder */
    remainder?: (IVar|null);
}
// ...
export interface IEMap {

    /** EMap kvs */
    kvs?: (IKeyValuePair[]|null);

    /** EMap locallyFree */
    locallyFree?: (Uint8Array|null);

    /** EMap connectiveUsed */
    connectiveUsed?: (boolean|null);

    /** EMap remainder */
    remainder?: (IVar|null);
}
// ...
export interface IEMethod {

    /** EMethod methodName */
    methodName?: (string|null);

    /** EMethod target */
    target?: (IPar|null);

    /** EMethod arguments */
    "arguments"?: (IPar[]|null);

    /** EMethod locallyFree */
    locallyFree?: (Uint8Array|null);

    /** EMethod connectiveUsed */
    connectiveUsed?: (boolean|null);
}
// ...
export interface IEVar {

    /** EVar v */
    v?: (IVar|null);
}
// ...
export interface IENot {

    /** ENot p */
    p?: (IPar|null);
}
// ...
export interface IENeg {

    /** ENeg p */
    p?: (IPar|null);
}
// ...
export interface IEMult {

    /** EMult p1 */
    p1?: (IPar|null);

    /** EMult p2 */
    p2?: (IPar|null);
}
// ...
export interface IEDiv {

    /** EDiv p1 */
    p1?: (IPar|null);

    /** EDiv p2 */
    p2?: (IPar|null);
}
// ...
export interface IEPlus {

    /** EPlus p1 */
    p1?: (IPar|null);

    /** EPlus p2 */
    p2?: (IPar|null);
}
// ...
export interface IEMinus {

    /** EMinus p1 */
    p1?: (IPar|null);

    /** EMinus p2 */
    p2?: (IPar|null);
}
// ...
export interface IELt {

    /** ELt p1 */
    p1?: (IPar|null);

    /** ELt p2 */
    p2?: (IPar|null);
}
// ...
export interface IELte {

    /** ELte p1 */
    p1?: (IPar|null);

    /** ELte p2 */
    p2?: (IPar|null);
}
// ...
export interface IEGt {

    /** EGt p1 */
    p1?: (IPar|null);

    /** EGt p2 */
    p2?: (IPar|null);
}
// ...
export interface IEGte {

    /** EGte p1 */
    p1?: (IPar|null);

    /** EGte p2 */
    p2?: (IPar|null);
}
// ...
export interface IEEq {

    /** EEq p1 */
    p1?: (IPar|null);

    /** EEq p2 */
    p2?: (IPar|null);
}
// ...
export interface IENeq {

    /** ENeq p1 */
    p1?: (IPar|null);

    /** ENeq p2 */
    p2?: (IPar|null);
}
// ...
export interface IEAnd {

    /** EAnd p1 */
    p1?: (IPar|null);

    /** EAnd p2 */
    p2?: (IPar|null);
}
// ...
export interface IEOr {

    /** EOr p1 */
    p1?: (IPar|null);

    /** EOr p2 */
    p2?: (IPar|null);
}
// ...
export interface IConnective {

    /** Connective connAndBody */
    connAndBody?: (IConnectiveBody|null);

    /** Connective connOrBody */
    connOrBody?: (IConnectiveBody|null);

    /** Connective connNotBody */
    connNotBody?: (IPar|null);

    /** Connective varRefBody */
    varRefBody?: (IVarRef|null);

    /** Connective connBool */
    connBool?: (boolean|null);

    /** Connective connInt */
    connInt?: (boolean|null);

    /** Connective connString */
    connString?: (boolean|null);

    /** Connective connUri */
    connUri?: (boolean|null);

    /** Connective connByteArray */
    connByteArray?: (boolean|null);
}
// ...
export interface IDeployId {

    /** DeployId sig */
    sig?: (Uint8Array|null);
}
// ...
export interface IDeployerId {

    /** DeployerId publicKey */
    publicKey?: (Uint8Array|null);
}
// ...
export interface IGUnforgeable {

    /** GUnforgeable gPrivateBody */
    gPrivateBody?: (IGPrivate|null);

    /** GUnforgeable gDeployIdBody */
    gDeployIdBody?: (IGDeployId|null);

    /** GUnforgeable gDeployerIdBody */
    gDeployerIdBody?: (IGDeployerId|null);
}

/** Represents a GUnforgeable. */
export class GUnforgeable implements IGUnforgeable {

    /**
     * Constructs a new GUnforgeable.
     * @param [properties] Properties to set
     */
    constructor(properties?: IGUnforgeable);

    /** GUnforgeable gPrivateBody. */
    public gPrivateBody?: (IGPrivate|null);

    /** GUnforgeable gDeployIdBody. */
    public gDeployIdBody?: (IGDeployId|null);

    /** GUnforgeable gDeployerIdBody. */
    public gDeployerIdBody?: (IGDeployerId|null);

    /** GUnforgeable unfInstance. */
    public unfInstance?: ("gPrivateBody"|"gDeployIdBody"|"gDeployerIdBody");

    /**
     * Creates a new GUnforgeable instance using the specified properties.
     * @param [properties] Properties to set
     * @returns GUnforgeable instance
     */
    public static create(properties?: IGUnforgeable): GUnforgeable;

    /**
     * Encodes the specified GUnforgeable message. Does not implicitly {@link GUnforgeable.verify|verify} messages.
     * @param message GUnforgeable message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encode(message: IGUnforgeable, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Encodes the specified GUnforgeable message, length delimited. Does not implicitly {@link GUnforgeable.verify|verify} messages.
     * @param message GUnforgeable message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encodeDelimited(message: IGUnforgeable, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a GUnforgeable message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns GUnforgeable
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): GUnforgeable;

    /**
     * Decodes a GUnforgeable message from the specified reader or buffer, length delimited.
     * @param reader Reader or buffer to decode from
     * @returns GUnforgeable
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): GUnforgeable;

    /**
     * Verifies a GUnforgeable message.
     * @param message Plain object to verify
     * @returns `null` if valid, otherwise the reason why it is not
     */
    public static verify(message: { [k: string]: any }): (string|null);

    /**
     * Creates a GUnforgeable message from a plain object. Also converts values to their respective internal types.
     * @param object Plain object
     * @returns GUnforgeable
     */
    public static fromObject(object: { [k: string]: any }): GUnforgeable;

    /**
     * Creates a plain object from a GUnforgeable message. Also converts values to other types if specified.
     * @param message GUnforgeable
     * @param [options] Conversion options
     * @returns Plain object
     */
    public static toObject(message: GUnforgeable, options?: $protobuf.IConversionOptions): { [k: string]: any };

    /**
     * Converts this GUnforgeable to JSON.
     * @returns JSON object
     */
    public toJSON(): { [k: string]: any };
}

/** Properties of a GPrivate. */
export interface IGPrivate {

    /** GPrivate id */
    id?: (Uint8Array|null);
}
// ...
export interface IGDeployId {

    /** GDeployId sig */
    sig?: (Uint8Array|null);
}
// ...
export interface IGDeployerId {

    /** GDeployerId publicKey */
    publicKey?: (Uint8Array|null);
}